Move RequestedShipments into its own folder
