+++ src/components/Office/RequestedShipments/RequestedShipments.module.scss
@@ -1,6 +1,6 @@
-@import '../../shared/styles/basics';
-@import '../../shared/styles/mixins';
-@import '../../shared/styles/colors';
+@import 'src/shared/styles/basics';
+@import 'src/shared/styles/mixins';
+@import 'src/shared/styles/colors';
 
 .RequestedShipments {
   @include u-bg('white');
